+++ Gems/ROS2/Code/Source/Manipulation/JointsManipulationComponent.cpp
@@ -21,6 +21,147 @@
 
 namespace ROS2
 {
+    namespace Internal
+    {
+        void AddHingeJointInfo(const AZ::EntityComponentIdPair& idPair, const AZStd::string& jointName, ManipulationJoints& joints)
+        {
+            if (joints.find(jointName) != joints.end())
+            {
+                AZ_Assert(false, "Joint names in hierarchy need to be unique (%s is not)!", jointName.c_str());
+                return;
+            }
+            AZ_Printf("JointsManipulationEditorComponent", "Adding joint info for hinge joint %s\n", jointName.c_str());
+            JointInfo jointInfo;
+            jointInfo.m_isArticulation = false;
+            jointInfo.m_axis = static_cast<PhysX::ArticulationJointAxis>(0);
+            jointInfo.m_entityComponentIdPair = idPair;
+            joints[jointName] = jointInfo;
+        }
+
+        bool TryGetFreeArticulationAxis(const AZ::EntityId& entityId, PhysX::ArticulationJointAxis& axis)
+        {
+            for (AZ::u8 i = 0; i <= static_cast<AZ::u8>(PhysX::ArticulationJointAxis::Z); i++)
+            {
+                PhysX::ArticulationJointMotionType type = PhysX::ArticulationJointMotionType::Locked;
+                axis = static_cast<PhysX::ArticulationJointAxis>(i);
+                // Use bus to prevent compilation error without PhysX Articulation support.
+                PhysX::ArticulationJointRequestBus::EventResult(type, entityId, &PhysX::ArticulationJointRequests::GetMotion, axis);
+                if (type != PhysX::ArticulationJointMotionType::Locked)
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        void AddArticulationJointInfo(const AZ::EntityComponentIdPair& idPair, const AZStd::string& jointName, ManipulationJoints& joints)
+        {
+            PhysX::ArticulationJointAxis freeAxis;
+            bool hasFreeAxis = TryGetFreeArticulationAxis(idPair.GetEntityId(), freeAxis);
+            if (!hasFreeAxis)
+            { // Do not add a joint since it is a fixed one
+                AZ_Printf("JointsManipulationEditorComponent", "Articulation joint %s is fixed, skipping\n", jointName.c_str());
+                return;
+            }
+
+            if (joints.find(jointName) != joints.end())
+            {
+                AZ_Assert(false, "Joint names in hierarchy need to be unique (%s is not)!", jointName.c_str());
+                return;
+            }
+
+            AZ_Printf("JointsManipulationEditorComponent", "Adding joint info for articulation link %s\n", jointName.c_str());
+            JointInfo jointInfo;
+            jointInfo.m_isArticulation = true;
+            jointInfo.m_axis = freeAxis;
+            jointInfo.m_entityComponentIdPair = idPair;
+            joints[jointName] = jointInfo;
+        }
+
+        ManipulationJoints GetAllEntityHierarchyJoints(const AZ::EntityId& entityId)
+        { // Look for either Articulation Links or Hinge joints in entity hierarchy and collect them into a map.
+            // Determine kind of joints through presence of appropriate controller
+            bool supportsArticulation = false;
+            bool supportsClassicJoints = false;
+            JointsPositionControllerRequestBus::EventResult(
+                supportsArticulation, entityId, &JointsPositionControllerRequests::SupportsArticulation);
+            JointsPositionControllerRequestBus::EventResult(
+                supportsClassicJoints, entityId, &JointsPositionControllerRequests::SupportsClassicJoints);
+            ManipulationJoints manipulationJoints;
+            if (!supportsArticulation && !supportsClassicJoints)
+            {
+                AZ_Warning("JointsManipulationEditorComponent", false, "No suitable Position Controller Component in entity!");
+                return manipulationJoints;
+            }
+            if (supportsArticulation && supportsClassicJoints)
+            {
+                AZ_Warning(
+                    "JointsManipulationEditorComponent", false, "Cannot support both classic joint and articulations in one hierarchy");
+                return manipulationJoints;
+            }
+
+            // Get all descendants and iterate over joints
+            AZStd::vector<AZ::EntityId> descendants;
+            AZ::TransformBus::EventResult(descendants, entityId, &AZ::TransformInterface::GetEntityAndAllDescendants);
+            AZ_Warning(
+                "JointsManipulationEditorComponent", descendants.size() > 0, "Entity %s has no descendants!", entityId.ToString().c_str());
+            for (const AZ::EntityId& descendantID : descendants)
+            {
+                AZ::Entity* entity = nullptr;
+                AZ::ComponentApplicationBus::BroadcastResult(entity, &AZ::ComponentApplicationRequests::FindEntity, descendantID);
+                AZ_Assert(entity, "Unknown entity %s", descendantID.ToString().c_str());
+
+                // If there is a Frame Component, take joint name stored in it.
+                auto* frameComponent = Utils::GetGameOrEditorComponent<ROS2FrameComponent>(entity);
+                if (!frameComponent)
+                { // Frame Component is required for joints.
+                    continue;
+                }
+                const AZStd::string jointName(frameComponent->GetJointName().GetCStr());
+
+                auto* hingeComponent = Utils::GetGameOrEditorComponent<PhysX::HingeJointComponent>(entity);
+                auto* articulationComponent = Utils::GetGameOrEditorComponent<PhysX::ArticulationLinkComponent>(entity);
+                AZ_Warning(
+                    "JointsManipulationEditorComponent",
+                    hingeComponent && !supportsClassicJoints,
+                    "Found classic joints but the controller does not support them!");
+                AZ_Warning(
+                    "JointsManipulationEditorComponent",
+                    articulationComponent && !supportsArticulation,
+                    "Found articulations but the controller does not support them!");
+
+                // See if there is a Hinge Joint in the entity, add it to map.
+                if (supportsClassicJoints && hingeComponent)
+                {
+                    auto idPair = AZ::EntityComponentIdPair(hingeComponent->GetEntityId(), hingeComponent->GetId());
+                    Internal::AddHingeJointInfo(idPair, jointName, manipulationJoints);
+                }
+
+                // See if there is an Articulation Link in the entity, add it to map.
+                if (supportsArticulation && articulationComponent)
+                {
+                    auto idPair = AZ::EntityComponentIdPair(articulationComponent->GetEntityId(), articulationComponent->GetId());
+                    Internal::AddArticulationJointInfo(idPair, jointName, manipulationJoints);
+                }
+            }
+            return manipulationJoints;
+        }
+
+        void SetInitialPositions(ManipulationJoints& manipulationJoints, const AZStd::unordered_map<AZStd::string, float>& initialPositions)
+        {
+            // Set the initial / resting position to move to and keep.
+            for (const auto& [jointName, jointInfo] : manipulationJoints)
+            {
+                AZ_Warning(
+                    "JointsManipulationEditorComponent",
+                    initialPositions.find(jointName.c_str()) != initialPositions.end(),
+                    "No set initial position for joint %s",
+                    jointName.c_str());
+                manipulationJoints[jointName].m_restPosition = initialPositions.at(jointName);
+            }
+        }
+    } // namespace Internal
+
     JointsManipulationComponent::JointsManipulationComponent()
     {
     }
@@ -183,9 +324,14 @@ namespace ROS2
     {
         if (m_manipulationJoints.empty())
         {
-            AZ_Warning("JointsManipulationComponent", false, "No manipulation joints to handle!");
-            AZ::TickBus::Handler::BusDisconnect();
-            return;
+            m_manipulationJoints = Internal::GetAllEntityHierarchyJoints(GetEntityId());
+            Internal::SetInitialPositions(m_manipulationJoints, m_initialPositions);
+            if (m_manipulationJoints.empty())
+            {
+                AZ_Warning("JointsManipulationComponent", false, "No manipulation joints to handle!");
+                AZ::TickBus::Handler::BusDisconnect();
+                return;
+            }
         }
         m_jointStatePublisher->OnTick(deltaTime);
         MoveToSetPositions(deltaTime);
